# expected-oas/contextapi.openapi.json
{
  "components": {
    "schemas": {
...
      "ContextInfo": {
        "description": "User-visible context information",
        "properties": {
          "lifetime": {
            "description": "Remaining seconds for the context",
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "lifetime"
        ],
        "type": "object"
      },
      "ContextPatch": {
        "description": "Parameter changes desired by user in a PATCH request",
        "properties": {
          "lifetime": {
            "description": "Context lifetime timeout in seconds",
            "format": "uint64",
            "minimum": 0.0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "ResourceConstraints": {
        "description": "Properties a suitable machine should satisfy.",
        "properties": {
          "machine_id": {
            "description": "The machine's unique id. This field is NOT recommended, but may be needed for certain use cases.",
            "format": "uint16",
            "minimum": 0.0,
            "type": [
              "integer",
              "null"
            ]
          },
          "platform": {
            "description": "The platform. This field must always be set.",
            "type": "string"
          }
        },
        "required": [
          "platform"
        ],
        "type": "object"
      },
...
      "Rsd": {
        "description": "Resource setup description",
        "properties": {
          "machines": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ResourceConstraints"
            },
            "type": "object"
          }
        },
        "required": [
          "machines"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The REST-API to access the Hardware as a Service implementation.",
    "title": "HWaaS Context API",
    "version": "5.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/contexts": {
      "post": {
        "description": "Reserves machines satisfying the given RSD. A context identifier is returned which is used to access the context consisting of the reserved machines.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rsd"
              }
            }
          },
          "description": "Resource setup description",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "plain text"
          },
          "422": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "The RSD contained an unsatisfiable constraint"
          }
        },
        "summary": "Reserve a context"
      }
    },
    "/contexts/{ctx_id}": {
      "delete": {
        "description": "Deletes the context and frees all of its assigned resources.",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "202": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "plain text"
          },
          "404": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "The context was not found"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Something went wrong"
          }
        },
        "summary": "Deletes a context"
      },
      "get": {
        "description": "Get configuration of the context.",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextInfo"
                }
              }
            },
            "description": "Context configuration"
          },
          "404": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "The context was not found"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Something went wrong"
          }
        },
        "summary": "Get context configuration"
      },
      "patch": {
        "description": "Modify configuration of the context.",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextPatch"
              }
            }
          },
          "description": "Parameter changes desired by user in a PATCH request",
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "plain text"
          },
          "404": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "The context was not found"
          },
          "409": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Context configuration not acceptable"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Something went wrong"
          }
        },
        "summary": "Modifies a context"
      }
    },
    "/contexts/{ctx_id}/drives": {
      "get": {
        "description": "Get the list of all allocated drives",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "list of drives"
          }
        },
        "summary": "list drives",
        "tags": [
          "Drives API"
        ]
      }
    },
    "/contexts/{ctx_id}/drives/{drive_name}": {
      "delete": {
        "description": "Delete a drive. \nPay attention: currently there is no check if the drive is mounted at any machine!",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "description": "user's name of the drive",
            "in": "path",
            "name": "drive_name",
            "required": true,
            "schema": {
              "description": "user's name of the drive",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "no content"
          },
          "404": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "no such drive"
          },
          "410": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Drive is already gone"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Drive allocation failed"
          }
        },
        "summary": "delete drive",
        "tags": [
          "Drives API"
        ]
      },
      "get": {
        "description": "Probe if a specified drive exists",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "description": "user's name of the drive",
            "in": "path",
            "name": "drive_name",
            "required": true,
            "schema": {
              "description": "user's name of the drive",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "drive exists"
          },
          "404": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "drive does not exist"
          }
        },
        "summary": "get drive",
        "tags": [
          "Drives API"
        ]
      },
      "put": {
        "description": "Create a new drive from a known image",
        "parameters": [
          {
            "description": "Context access token",
            "in": "path",
            "name": "ctx_id",
            "required": true,
            "schema": {
              "description": "Context access token",
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "description": "user's name of the drive",
            "in": "path",
            "name": "drive_name",
            "required": true,
            "schema": {
              "description": "user's name of the drive",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "description": "id of the drive",
            "in": "query",
            "name": "image_hash",
            "required": true,
            "schema": {
              "description": "id of the drive",
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "no content"
          },
          "409": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "drive exists"
          },
          "422": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "base image not available"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "drive creation failed internally"
          }
        },
        "summary": "create drive",
        "tags": [
          "Drives API"
        ]
      }
    },
...
  }
}